Create CalcTable.MultiValueField
Calculated table that addresses the issue of having a field that contains multiple values that must be split up. 

In this example, the field [Lead Lawyer] in the 'Proposals and Pitches' table has multiple values (lawyer names) for many records, and each value is delimited by the characters ";#". The resulting table pulls in [Id] and [Name] from the source table, and adds the newly created [Lawyer] field that now has individual values rather than multiple.