Create Measure.AES
Aggregate Engagement Score (AES): A relatively simple lead scoring formula that takes individual marketing activities (as tracked in a CRM system) and assigns an arbitrary weight to generate a summary score of a single contact's activities relative to other contacts in the CRM database. Positive activities are added, negative activities subtracted. Note that the email open statistic here ([AESopened]) is weighted very low relative to other activities, due to the recent updates in email open reporting.

In this example, we have created 16 independent measures that sum the number of activities for each of the 16 activity categories. The number of activities could be larger or smaller depending on the client/project. The weights are also dependent on the client/project. You will notice there is one measure that is subtracted from the overall score - [AESoptout]. This activity relates to a contact that opts out of future marketing communications, so that represents a negative impact on the overall AES.