Create CalcColumn.CountDuplicates
A calculated column that returns the number of times the value for a particular field appears across all records in the table. 

Great way to count unique contact ID duplicates in a CRM contact table.

Column will return numeric values. Any value greater than 1 indicates a duplicate for that record exists in the table.

In this example, the calculated column is looking at the [Contact ID] field in the 'All Contacts' table to report back any contact ID duplicates.

Note that this will NOT work as a measure...calculated column only.