Create Table.DimDate
M Code that creates a table with sequential dates running from 2015 through 2030. Includes several different ways of displaying dates including abbreviated names, Quarters, week and month sorting, and more.

This table is a MUST for any time-bound analyses you will be running.

This has to be entered as M Code and not as a measure.