Create CalcColumn.JobTitleGroups
A calculated column that creates a roll up of job titles based on a Job Title field in a CRM and assigns each title an arbitrary job title group. Titles are found by one of two mechanisms:

SEARCH: looks for defined text within overall text string
CONTAINSSTRINGEXACT: looks for an exact text string match

In this example, job titles are collected from the [Job Title] field of a contact table into one of five groups: Senior Leadership, Business Professionals, Legal Counsel, No Job Title and Other. Both English and French-Canadian variants of popular job titles in a legal setting are included.  

An example of the SEARCH function is SEARCH("Director".... - this will search the Job Title field for the word "director" and any job title that has that word will be assigned to the "Business Professionals" Group. 

However, for the example of CONTAINSSTRINGEXACT, let's look at CONTAINSSTRINGEXACT("COO"... this will look for records that ONLY have COO in the field, not words like "coordinator", which would otherwise be caught by the SEARCH function and grouped incorrectly.