Create Measure.CountRowsFilterByString
Trying to count the number of rows that contain a specific text string? This measure will help you do that. In the example, we are counting the number of rows in a table that contain the word "Manager" in a particular COLUMN. This will count any field in that column that has the word "manager" somewhere within.